Module Treemap

This implement a map from keys that are in a tree relationship (e.g. a sets of paths) to a lattice.

It is so that after n calls to refine ki li, then a map k is mapped to the intersection of all the li for which there was a refine ki li call, and for which ki is a parent of k.

module type Key = sig ... end

module Make_no_empty (Key : Key) : sig ... end

More efficient version, but without empty.

module Make (Key : Key) : sig ... end