Module type Treemap.Key

type t

val nearest_common_ancestor : t -> t -> t

Nearest common ancestor between two elements in the tree

val is_prefix : t -> t -> bool

is_prefix a b is true if a an ancestor of b

val equal : t -> t -> bool

val pretty : Format.formatter -> t -> unit